fix: update currency schema from float to int

+++ booking/internal/market/dto/dto.go
@@ -1,13 +1,13 @@
 package dto
 
 type TradeInput struct {
-	OrderID       string  `json:"order_id"`
-	InvestorID    string  `json:"investor_id"`
-	AssetID       string  `json:"asset_id"`
-	CurrentShares int     `json:"current_shares"`
-	Shares        int     `json:"shares"`
-	Price         float64 `json:"price"`
-	OrderType     string  `json:"order_type"`
+	OrderID       string `json:"order_id"`
+	InvestorID    string `json:"investor_id"`
+	AssetID       string `json:"asset_id"`
+	CurrentShares int    `json:"current_shares"`
+	Shares        int    `json:"shares"`
+	Price         int    `json:"price"`
+	OrderType     string `json:"order_type"`
 }
 
 type OrderOutput struct {
@@ -22,10 +22,10 @@ type OrderOutput struct {
 }
 
 type TransactionOutput struct {
-	TransactionID string  `json:"transaction_id"`
-	BuyerID       string  `json:"buyer_id"`
-	SellerID      string  `json:"seller_id"`
-	AssetID       string  `json:"asset_id"`
-	Price         float64 `json:"price"`
-	Shares        int     `json:"shares"`
+	TransactionID string `json:"transaction_id"`
+	BuyerID       string `json:"buyer_id"`
+	SellerID      string `json:"seller_id"`
+	AssetID       string `json:"asset_id"`
+	Price         int    `json:"price"`
+	Shares        int    `json:"shares"`
 }
